requirements.txt
albumentations==1.1.0
Cython==0.29.24
future==0.18.2
loguru==0.5.3
numpy==1.21.4
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.58
opencv-python-headless==4.5.5.62
Pillow==8.4.0
segmentation-models-pytorch==0.2.1
torch==1.8.1
torchvision==0.9.1
tqdm==4.62.3
matplotlib==3.4.3
pandas==1.3.4
